// src/savvy/redux/types/H01_Manager/ManagerType.ts
/**
 * @description MerchManager reducer
 * @author Xupeng Bao
 * @date 2021-05-24
 */

export enum ManagerType {
    MANAGER_SELECTED_DATE = 'manager_selectedDate',
    MANAGER_SEARCH_TEXT = 'manager_searchText',
    MANAGER_SCHEDULE_DATE = 'manager_scheduleDate',
    MANAGER_IS_COMPLETED = 'manager_isCompleted',
    MANAGER_IS_CANCELING = 'manager_isCanceling',
    MANAGER_IS_CANCELED = 'manager_IsCanceled',
    MANAGER_VISIT_LIST_ID = 'manager_visitListId',
    MANAGER_IS_CANCEL_BTN_CLICKED = 'manager_isCancelBtnClicked',
    MANAGER_MY_TEAM_DATA = 'manager_myTeamData',
    MANAGER_EMPLOYEE_LIST = 'manager_employeeList',
    MANAGER_CUSTOMER_LIST = 'manager_customerList',
    MANAGER_EMPLOYEE_LIST_FOR_REASSIGN = 'manager_employeeListForReassign',
    MANAGER_RNS_EMPLOYEE_LIST = 'manager_RNS_employeeList',
    MANAGER_RNS_CUSTOMER_LIST = 'manager_RNS_customerList',
    MANAGER_LINE_CODE_MAP = 'manager_line_code_map',
    MANAGER_NOTIFICATION_DATA = 'manager_notification_data',
    MANAGER_AVATAR_REFRESH = 'manager_avatar_refresh',
    MANAGER_COMPLETED_SVG_LIST = 'manager_completedSVGList',
    MANAGER_PUBLISHED_SVG_LIST = 'manager_publishedSVGList',
    MANAGER_LOCATION_INFO = 'manager_locationInfo',
    MANAGER_IS_IN_CANADA = 'manager_isInCanada'
}